requirements.txt
selenium
requests
boto3
warrant
pytest
mail-parser
beautifulsoup4
lxml
pylint
pytest-repeat
pyyaml
flask
pydantic
pysftp
